fix(web): share route distance formatting
